+++ src/test/converter2/issues/gh2290.ts
@@ -1,0 +1,14 @@
+/** Int comment */
+export interface CallSignature {
+    /** Sig comment */
+    (x: string): void;
+}
+
+/** Int comment */
+export interface CallSignature2 {
+    /** Sig comment */
+    (x: string): void;
+
+    /** Prop comment */
+    prop: 123;
+}